backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
pyjwt
python-multipart
pdfplumber
pytesseract
spacy
nltk
scikit-learn
pandas
numpy
python-dotenv
alembic
pydantic[email]
jinja2
python-jose[cryptography]
google-generativeai
google-genai
